Added the Calender.html file. It's displaying the December 2022 calendar
The styling used is internal CSS. The Flex-shrink property specifies how the item will shrink relative to the rest of the flexible items inside the same container. The Flex basis property defines the size of the Flex-item along the main axis of the flex container . The Flex grow property specifies how much the item will grow relative to the rest of the items inside the same container.                                                          
NOTE: If the element is not a flexible item ,the flex property has no effect. The default value is 0.